Moved @tryghost/bookshelf-repository to recommendations package
fix https://linear.app/ghost/issue/ENG-2095/bookshelf-repository

- this commit moves the bookshelf repository to the recommendations
  folder, which is the only place it is used and helps in the efforts to
  de-package the monorepo
