--- ghost/bookshelf-repository/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@tryghost/bookshelf-repository",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/ghost/bookshelf-repository",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "build/index.js",
-  "types": "build/index.d.ts",
-  "scripts": {
-    "build": "yarn build:ts",
-    "build:ts": "tsc",
-    "test:unit": "NODE_ENV=testing c8 --src src --all --check-coverage --100 --reporter text --reporter cobertura -- mocha --reporter dot -r ts-node/register './test/**/*.test.ts'",
-    "test": "yarn test:types && yarn test:unit",
-    "test:types": "tsc --noEmit",
-    "lint:code": "eslint src/ --ext .ts --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .ts --cache"
-  },
-  "files": [
-    "build"
-  ],
-  "devDependencies": {
-    "c8": "7.14.0",
-    "mocha": "10.2.0",
-    "sinon": "15.2.0",
-    "@tryghost/nql": "0.12.7"
-  },
-  "dependencies": {
-    "@tryghost/mongo-utils": "0.6.2",
-    "knex": "2.4.2"
-  }
-}
