--- ghost/bookshelf-repository/src/libraries.d.ts
@@ -1,3 +1,0 @@
-declare module '@tryghost/mongo-utils';
-declare module '@tryghost/errors';
-declare module '@tryghost/nql';
