+++ ghost/recommendations/package.json
@@ -22,6 +22,7 @@
     "build"
   ],
   "devDependencies": {
+    "@tryghost/nql": "0.12.7",
     "@types/node": "20.14.8",
     "c8": "8.0.1",
     "mocha": "10.2.0",
@@ -33,7 +34,8 @@
     "@tryghost/tpl": "0.1.32",
     "@tryghost/errors": "1.3.5",
     "@tryghost/in-memory-repository": "0.0.0",
-    "@tryghost/bookshelf-repository": "0.0.0",
-    "@tryghost/logging": "2.4.18"
+    "@tryghost/logging": "2.4.18",
+    "@tryghost/mongo-utils": "0.6.2",
+    "knex": "2.4.2"
   }
 }
