+++ ghost/recommendations/src/libraries.d.ts
@@ -1,3 +1,5 @@
 declare module '@tryghost/errors';
 declare module '@tryghost/tpl';
 declare module '@tryghost/logging';
+declare module '@tryghost/mongo-utils';
+declare module '@tryghost/nql';
