+++ ghost/recommendations/test/BookshelfRepository.test.ts
@@ -1,7 +1,7 @@
 import nql from '@tryghost/nql';
 import assert from 'assert/strict';
 import {Knex} from 'knex';
-import {BookshelfRepository, ModelClass, ModelInstance} from '../src/index';
+import {BookshelfRepository, ModelClass, ModelInstance} from '../src/BookshelfRepository';
 
 type SimpleEntity = {
     id: string;
